requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.5.0
    # via starlette
asgiref==3.5.1
    # via uvicorn
click==8.1.3
    # via uvicorn
dnspython==2.2.1
    # via pymongo
fastapi==0.77.1
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
idna==3.3
    # via anyio
pydantic==1.9.0
    # via fastapi
pymongo[srv]==4.1.1
    # via -r requirements.in
python-decouple==3.6
    # via -r requirements.in
sniffio==1.2.0
    # via anyio
starlette==0.19.1
    # via fastapi
typing-extensions==4.2.0
    # via pydantic
uvicorn==0.17.6
    # via -r requirements.in
